+++ src/routes/api/v2/journal/+server.ts
@@ -6,14 +6,14 @@ import { checkAuth } from '$lib/server/utils';
 
 import type { RequestHandler } from './$types';
 
-export const GET = withRequestHandlerMiddleware<RequestHandler>(async ({ cookies }) => {
-  const { uuid } = await checkAuth(cookies);
+export const GET = withRequestHandlerMiddleware<RequestHandler>(async ({ cookies, request }) => {
+  const { uuid } = await checkAuth(cookies, request);
   const result = await getJournal({ start: 0 }, uuid);
   return json(result);
 });
 
 export const POST = withRequestHandlerMiddleware<RequestHandler>(async ({ request, cookies }) => {
-  const { uuid } = await checkAuth(cookies);
+  const { uuid } = await checkAuth(cookies, request);
   const data = await request.json();
   const result = await postManyJournal(data, uuid);
   return json(result);